pyproject.toml
[project]
name = "pos-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "authx>=1.4.3",
    "click>=8.3.1",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "jose>=1.0.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
